+++ invenapp/src/app/dashboard/components/lista-usuario/lista-usuario.component.ts
@@ -4,6 +4,7 @@ import { DbHandlerService } from "../../services/db-handler.service";
 import { FormBuilder, FormGroup, FormControl, Validators  } from "@angular/forms";
 import { Router } from "@angular/router";
 import { forkJoin } from "rxjs";
+import { faTrashAlt, faFilePdf, faEdit, faEye } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
   selector: 'app-lista-usuario',
@@ -12,15 +13,21 @@ import { forkJoin } from "rxjs";
 })
 export class ListaUsuarioComponent implements OnInit {
 
+  faTrash = faTrashAlt;
+  faEye = faEye;
+  faPdf = faFilePdf;
+  faEdit = faEdit;
+
   isEmpty: boolean;
   endpoint: string;
   name: string;
   title: string;
   addText: string;
-  fields: string[];
-  values: string[];
+  fields: any[];
+  values: any[];
 
   constructor(
+    private auth: AuthService,
     private dbHandler: DbHandlerService,
     private router: Router,
     private fb: FormBuilder
@@ -30,16 +37,36 @@ export class ListaUsuarioComponent implements OnInit {
 
   ngOnInit() {
     this.initComponent('/users', 'Lista de Usuarios', 'Agregar Usuario', 'users')
-    this.isEmpty = true;
     this.initForm();
+    this.isEmpty = true;
+     let auxfields = this.dbHandler.getLocal(this.name + 'Fields');
+     let auxValues = this.dbHandler.getLocal(this.name + 'Values');
 
-    this.fields = this.dbHandler.getLocal(this.name + 'Fields');
-    this.values = this.dbHandler.getLocal(this.name + 'Values');
+    this.fields = [
+      'Id',
+      'Nombre',
+      'Usuario',
+      'Correo',
+      'Tipo',
+      ]
+
+    this.values = [];
+
+    auxValues.forEach(value => {
+      let aux = [
+        value._id,
+        value.name,
+        value.username,
+        value.mail,
+        value.type,
+      ]
+      this.values.push(aux)
+    });
+
     if(this.values.length){
       this.isEmpty = false;
     }
   }
-
 
   initComponent(endpoint, title, addText, name) {
     this.endpoint = endpoint;
@@ -48,24 +75,63 @@ export class ListaUsuarioComponent implements OnInit {
     this.name = name;
   }
 
+  isAdmin: boolean;
+  isSuperAdmin: boolean;
+  isVendedor: boolean;
 
-  deleteItem(event, item) {
+  deleteItem(event, index) {
+    let auxValues = this.dbHandler.getLocal(this.name + 'Values');
+    let item = auxValues[index];
     var myEnd = this.endpoint;
+    let type = this.auth.getType();
+    this.isAdmin = (type === 'Admin');
+		this.isVendedor = (type === 'Vendedor');
+		this.isSuperAdmin = (type === 'SuperAdmin');
+    
+    //Autorizacion basada en roles. Modificar eventualmente a basada en reglas
+    console.log(this.isSuperAdmin)
+    console.log(this.isAdmin)
+    if(!(this.isSuperAdmin)){
+          this.closeConfirm();
+          let errorMsg = 'Usuario no autorizado';
+          this.openError(errorMsg)
+    } else{
 
-    this.dbHandler.deleteSomething(item._id, myEnd)
-      .subscribe((data: any) => {   // data is already a JSON object
+      this.dbHandler.deleteSomething(item._id, myEnd)
+      .subscribe((data: any) => { 
+        this.closeConfirm();
         if(!data.status){
           let errorMsg = data.msg;
           this.openError(errorMsg)
         } else{
-
         this.dbHandler.actualizar();
         }
       });
+    }
+  }
+
+  deletedItem: any;
+  confirmDelete(event, item){
+    this.deletedItem = item;
+    this.openConfirm();
+  }
+
+  showConfirm: {};
+
+  openConfirm(){
+    this.showConfirm = {
+        confirmAct: true,
+      }
+  }
+
+  closeConfirm(){
+    this.showConfirm = {
+        confirmAct: false,
+      }
   }
 
     openUpdate(event, item) {  
-      this.router.navigateByUrl('/actualizar/usuario');
+      this.router.navigateByUrl('/actualizar/usuario/'+item);
     }
 
    agregar() {
